pyproject.toml
[tool.poetry]
name = "identity"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "identity", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
fastapi = "^0.95.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.8"}
asyncpg = "^0.27.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-mock = "^3.10.0"
alembic = "^1.10.2"
aiosqlite = "^0.18.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
